Updates for compatibility and migration script.
Updated .bashrc to fix a shell prompt bug which only appears on certain
terminals, and also fixed .vimrc to work on systems without a p4 binary
available (not sure why this didn't break before, but it's not worth
looking into too deeply). Also added simple Python script to migrate
files from an initial clone into a subdir to be within the home dir.