Added better support for Windows home dirs.
Added snippet to .vimrc which detects Windows (using has() with "win32" and
"win16") and updates runtimepath to include ".vim" within the user's home
directory if both $HOMEDRIVE and $HOMEPATH are set. This appears to work
on at least my Windows machine at work - this may need updating at a later
stage as I try it on other Windows machines.