Moved grep options from env var to alias.
Use of the GREP_OPTIONS environment variable causes various problems in
build systems - for example the QNX build process and the Linux kernel.
Using an alias means that at least build systems which include the full
path of the grep binary will work, although it's entirely possible that
some scripts may still latch on to the alias.