Tmux auto-attach and Vim on iPad SSH
Amends .bashrc to automatically connect to any active tmux session on
login if the .localbashrc contains TMUX_AUTO_ATTACH=1. Also, some tweaks
to the tmux configuration to use CTRL-j as the escape (to match when I
use for screen, as it doesn't conflict with most curses apps) and some
other improvements. Finally, if the terminal string is "xterm-ios", Vim
remaps backtick to act as ESC as when using a bluetooth keyboard on the
iPad the ESC key is stolen for use as the HOME key. Note that this
terminal type isn't used by default on iSSH, but it can be configured to
do so.