Replace Vim pyflakes with syntastic
This commit removes the Vim pyflakes plugin as it's been replaced with
the syntastic Vim plugin which supports multiple languages. The previous
commit merged in the syntastic subtree, hence all that's left is to
remove pyflakes with this commit. Also added support for a .localvimrc
file, required for site-specific configuration changes.