Added settings for virtualenvwrapper
This commit adds a section to .bashrc to set the required environment
for virtualenvwrapper, storing the environments in `~/virtualenvs` (note
the slight difference from the suggested default `~/.virtualenvs`) and
the project directories under `~/src`.

Currently the only supported location for virtualenvwrapper to be
installed is in `/usr/local`, presuming that it'll always be installed
by `pip`, but this may need to be made more flexible in time.