Session logging, Vim changes, Git ctags hooks
This commit adds a bash function `logsession()` which is a simple
wrapper around the `script` binary for easily creating timestamped log
files of interactive console sessions. The contents of `.vimrc` have
been rationalised slightly and the `:Hardcopy` function now supports
ranges. Git has had a template directory added which maintains a global
ctags file inside the `.git` directory of the repository, updating it
every time the contents of the repository are changed. This location is
read by Vim whenever it loads the fugitive extension, although this
won't work if using the `-t` option on the command-line. Also, Vim
should now to proper comment continuation for Doxygen comments. Finally,
a change to `.tmux.conf` makes `new-session` the default command - this
allows the command `tmux attach` to attach to an existing session if
found, and create a new session if not.