Added completion to "templ" plugin
Added template name completion to my simple `templ` Vim plugin.
Currently the completion is quite simplistic and only works for
tail-completion (i.e. hitting TAB at the end of a prefix) - it won't
work correctly for completion when there are any trailing characters
after the cursor. This is an unusual use-case and quite fiddly to
implement properly, so I don't think it's worth the effort.