Added keychain support to .bashrc
This commit adds code to .bashrc which looks for an installation of
keychain on the system and, if found, invokes it with all private keys
matching `~/.ssh/id_rsa*` (filtering off those ending `.pub`). On a
system without keychain, this code is silently skipped.

The consequence of this is that on first login when the persistent
ssh-agent isn't running (e.g. after a reboot), the user will be prompted
for the passphrases for all their keys. Hitting CTRL+C will get past the
prompt, but the user will be prompted again next login.